Store CIT/COT as float (sub-degree resolution for CE)
CE divides by (COT-CIT); whole-degree storage was the noise floor. Fields ->
float, sensor writes drop the int cast, debug prints %.1f. j1939 doesn't TX
these, control loops don't read them.

+++ include/AppData.h
@@ -11,13 +11,13 @@ struct AppData {
     uint16_t actuatorMotorLoad;
     uint8_t actuatorStatus;
     uint16_t compressorOutputPressureHpaa;
-    int16_t compressorInputTempC;
+    float compressorInputTempC;
     uint16_t compressorInputPressureHpaa;
     uint16_t turbineInputPressureHpa;
     float turbineInputVoltage;
     float tipZeroVoltage;
     int16_t turbineInletTempC;
-    int16_t compressorOutputTempC;
+    float compressorOutputTempC;
     int16_t turbineOutletTempC;
     int16_t oilTempC;
     uint16_t oilPressureHpa;
